% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\title{Query data for standard data packet}
\usage{
query_data(dsn, query, demo_names = c("id", "term", "subject", "subject_long",
  "course_number", "age", "matr_goal", "ed_level", "ethnicity", "enroll_status",
  "inst_mode", "language", "gender", "grade"))
}
\arguments{
\item{dsn}{A data source name for the archive side of the LRDB. See the
readme on how to set up a data source name.}

\item{query}{A sql query in the form of a character vector or length one
the query must pull student id, term (e.g, 1139), subject (e.g., 'PSYC'),
and long subject ('Psychology')}

\item{demo_names}{A character vector of column names corresponding with
the order of columns in your sql query. The column names must be from
a standard list - each column name corresponding with currently supported
data (see details) and must contain 'id','term','subject', and 'subject_long'}
}
\description{
A function for querying demographic data for standard college
data packets. Returns an object with class data.frame and lrdataset
}
\details{
In order for this function to yield the necessary data, the user
must pass a sql query (param query) for the archive LRDB and a vector of
column names (param demo_names) corresponding with the columns of the query.
For example, if the first column of the query selected student id, then the
first column name provided would be 'id'. The function will only support
certain types of query data (e.g., student id, age, etc.), and each of these
data types has a standard column name that must be supplied in the same order
as the columns in the query. Each column name is listed below with associated
query data.

'id' - student id

'term' - 4-digit term identifier

'subject' - 2-5 letter course prefix (e.g., 'PSYC')

'long_subject' - full subject description (e.g, 'Psychology')

'course_number' - course number, up to 3 digits (e.g., 300)

'age' - student age during a given term

'matr_goal' - matriculation goal description (e.g.'Transfer to a 4-year')

'ed_level' - education level description (e.g., 'Completed GED')

'ethnicity' - student race

'inst_mode' - instructional mode description (e.g., 'lecture')

'language' - student's primary language

'gender' - student gender (m,f,u)

'grade' - official course grade (e.g., 'A','B','C', etc.)


Do not modify or recode the data in any way (e.g., with a case when). The
column names within the LRDB have not been provided here to keep the
structure of the database secure. A standard query can be provided upon
request. Contact \email{meinzp@crc.losrios.edu}.
}
\examples{
\dontrun{
query_data(dsn = 'datasourcename',
           query = 'select id, term, subject, subject_descr...',
           demo_names = c('id','term','subject','subject_long'))
           }
}

