% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{disag_hc_plot}
\alias{disag_hc_plot}
\alias{disag_scs_plot}
\alias{plot_headcounts}
\title{Generate a summary demographic plot}
\usage{
disag_hc_plot(data, demo_col, path, x_axis_lab = "", title = "")

disag_scs_plot(data, demo_col, path, x_axis_lab = "", title = "")

plot_headcounts(data, path, title = "", undup = TRUE)
}
\arguments{
\item{data}{A lrdataset object generated from the query_gen function.}

\item{demo_col}{The name of a demographic variable in the lrdataset object.
This demographic variable will be used to disaggregate.}

\item{path}{A file path to where you would like to save your plot.}

\item{x_axis_lab}{The x-axis label.}

\item{title}{The plot title}
}
\description{
Takes a lrdataset object (generated from the query_data function) and
produces a plot for the past five years disaggregated by a
demographic variable. disag_hc_plot produces a disaggregated demographic
plot, disag_scs_plot produces a disaggregated success plot, and
plot_headcounts produces a headcount (duplicated or unduplicated) plot.
}
\details{
You must use a lrdataset object with this function. The purpose of this
restriction is to ensure that the dataset has only standard data and column
names
}
\section{Functions}{
\itemize{
\item \code{disag_scs_plot}: Produce a Success Rate Plot

\item \code{plot_headcounts}: Produce a Headcount Plot
}}
\examples{
\dontrun{disag_hc_plot(data, 'gender', 'User/documents/ex.jpg','Gender',
'xaxislabel')}
}

